adds IRepair interface and adds to CarDealership with repair method and tests - PASSED. Fixes incorrect var type in IDamage and associated files

+++ src/main/java/behaviours/IDamage.java
@@ -2,5 +2,5 @@ package behaviours;
 
 public interface IDamage {
 
-    void damageCost(int cost);
+    void damageCost(double cost);
 }
